+++ assets/appFramework/vue/generate/component/src/{GEN_NAME_SNAKE}.js
@@ -4,10 +4,9 @@
 import Vue from "vue";
 import Component from "vue-class-component";
 import "./{GEN_NAME_SNAKE}.css";
-import importedTemplate from "./{GEN_NAME_SNAKE}.vue";
 
 @Component({
-    "template": importedTemplate
+    "template": "./{GEN_NAME_SNAKE}.vue"
 })
 export class {GEN_NAME_PASCAL} extends Vue {
 }