+++ assets/appFramework/vue/generate/component/src/{GEN_NAME_SNAKE}.ts
@@ -4,10 +4,9 @@
 import {SYNTHETIC_IMPORT}Vue from "vue";
 import Component from "vue-class-component";
 import "./{GEN_NAME_SNAKE}.css";
-import {SYNTHETIC_IMPORT}importedTemplate from "./{GEN_NAME_SNAKE}.vue";
 
 @Component({
-    template: importedTemplate
+    template: "./{GEN_NAME_SNAKE}.vue"
 })
 export class {GEN_NAME_PASCAL} extends Vue {
 }